completed Project Controller

+++ src/main/java/com/example/sdp3/Service/UserLicenseService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class UserLicenseService {
@@ -14,25 +15,9 @@ public class UserLicenseService {
     @Autowired
     UserLicenseRepository userLicenseRepository;
 
-    public List<UserLicense> getAllLicenses(Long id)
+    public List<UserLicense> getAllUserLicenses(Long id)
     {
-        try
-        {
-            List<UserLicense> userLicenses = new ArrayList<>();
-            List<UserLicense> data = (List<UserLicense>)userLicenseRepository.findAll();
-            for(UserLicense userLicense: data)
-            {
-                if (userLicense.getUser_id() == id)
-                {
-                    userLicenses.add(userLicense);
-                }
-            }
-            return userLicenses;
-        }
-        catch(Exception e)
-        {
-            return null;
-        }
+        return userLicenseRepository.findUserLicenseByUser_id(id).orElseThrow(() -> new IllegalStateException("No License found"));
     }
 
     public void createUserLicense(Long id , UserLicense userLicense)
@@ -50,39 +35,29 @@ public class UserLicenseService {
 
     public void deleteUserLicense(Long id)
     {
-        try
-        {
-            userLicenseRepository.deleteById(id);
-        }
-        catch(Exception e)
-        {
-            System.out.println(e);
-        }
+       boolean b = userLicenseRepository.existsById(id);
+       if (!b) {
+           throw new IllegalStateException("License Doesn't Exists");
+       }
+
+       userLicenseRepository.deleteById(id);
     }
 
     public UserLicense getUserLicenseById(Long id)
     {
-        try {
-            return userLicenseRepository.findById(id).get();
-        }
-        catch(Exception e)
-        {
-           return null;
-        }
+        return userLicenseRepository.findById(id).orElseThrow(() -> new IllegalStateException("License Not found!"));
     }
 
 
     public void updateUserLicense(UserLicense userLicense)
     {
-        try
-        {
-            userLicense.setUser_id(userLicense.getId());
+       UserLicense userLicense1 = userLicenseRepository.findById(userLicense.getId()).orElseThrow(() -> new IllegalStateException("License Not Found!"));
+
+        if (userLicense.getLicense_name().length() > 0 && !Objects.equals(userLicense1.getLicense_name(), userLicense.getLicense_name()) ) {
             userLicenseRepository.save(userLicense);
+
         }
-        catch (Exception e)
-        {
-            System.out.println("Error Occured while updating the User Profile");
-        }
+
 
     }
